Reorganize Lua tests by module file
Mirror the source-file split in the test layout so each module's
behavior lives next to the module it covers and runtime/plugin tests
stay focused on their respective concerns.

- module_commands_test.go: command-parsing success and rejection cases
  (was TestPluginInitSuccessCases / parsing portion of
  TestPluginInitRejectsInvalidInputs / TestPluginInitRejectsUnsupportedCommandFields).
- module_logger_test.go: hive.log invocation
  (was TestHiveLogFunctionsAreInvocable in runtime_test.go).
- runtime_test.go: keeps validateModuleName + require traversal and
  picks up TestLoadEntrypointRejectsInvalidArity, split out of the
  rejection table.
- plugin_test.go: pure lifecycle (Available, double-Init, partial-init
  failure) plus the shared NewConfigPlugin helper.

Tests stay integration-style — driven through Plugin.Init / NewRuntime
so wiring bugs still get caught — and there is no module_info_test.go
since the metadata surface is already asserted from
TestCommandsModuleSuccessCases.
